+++ app/dependencies.py
@@ -1,7 +1,7 @@
 # 2024/2/6
 # zhangzhong
 
-from app_refactor.database.proxy import DatabaseProxy
+from app.database.proxy import DatabaseProxy
 
 
 def database_proxy() -> DatabaseProxy:
