+++ app/main.py
@@ -1,8 +1,8 @@
 from fastapi import FastAPI
-from app_refactor.routers import character
+from app.routers import character
 from fastapi.middleware.cors import CORSMiddleware
 import uvicorn
-from app_refactor.common.conf import conf
+from app.common.conf import conf
 
 app = FastAPI()
 app.include_router(router=character.router)
